+++ backend/src/baserow/contrib/database/fields/models.py
@@ -123,6 +123,10 @@ class Field(
         result = []
         for field_dependency in self.dependants.select_related("dependant").all():
             dependant_field = field_cache.lookup_specific(field_dependency.dependant)
+            if dependant_field is None:
+                # If somehow the dependant is trashed it will be None. We can't really
+                # trigger any updates for it so ignore it.
+                continue
             dependant_field_type = field_type_registry.get_by_model(dependant_field)
             if field_dependency.via is not None:
                 via_path_to_starting_table = (